backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
finnhub-python==2.4.18

# Testing and CI
pytest==8.3.2
pytest-django==4.9.0
coverage==7.5.4

# Optional: for API mocking and response checks
responses==0.25.3
